Port TestLocalDateTime to scalatest
